refactor: 거래내역 그룹화/정렬 기능 함수화
